refactor E2EMappingIT to remove unnecessary bloat Move RandomIdGenerator MockBean to E2EMappingIT
